Update controller spec
